Tidy up block submission errors
