+++ go/common/errutil/errors_util.go
@@ -2,8 +2,12 @@ package errutil
 
 import (
 	"errors"
+	"fmt"
+	"strings"
 
 	"github.com/ethereum/go-ethereum"
+
+	gethcommon "github.com/ethereum/go-ethereum/common"
 )
 
 var (
@@ -11,4 +15,34 @@ var (
 	// we want to be able to catch both types in a single error-check.
 	ErrNotFound = ethereum.NotFound
 	ErrNoImpl   = errors.New("not implemented")
+
+	// Standard errors that can be returned from block submission
+
+	ErrBlockAlreadyProcessed = errors.New("block already processed")
+	ErrBlockAncestorNotFound = errors.New("block ancestor not found")
 )
+
+// BlockRejectError is used as a standard format for error response from enclave for block submission errors
+// The L1 Head hash tells the host what the enclave knows as the canonical chain head, so it can feed it the appropriate block.
+type BlockRejectError struct {
+	L1Head  gethcommon.Hash
+	Wrapped error
+}
+
+func (r BlockRejectError) Error() string {
+	head := "N/A"
+	if r.L1Head != (gethcommon.Hash{}) {
+		head = r.L1Head.String()
+	}
+	return fmt.Sprintf("%s l1Head=%s", r.Wrapped.Error(), head)
+}
+
+func (r BlockRejectError) Unwrap() error {
+	return r.Wrapped
+}
+
+// Is implementation supports the errors.Is() behaviour. The error being sent over the wire means we lose the typing
+// on the `Wrapped` error, but comparing the string helps in our use cases of standard exported errors above
+func (r BlockRejectError) Is(err error) bool {
+	return strings.Contains(r.Error(), err.Error()) || errors.Is(err, r.Wrapped)
+}
