+++ go/common/rpc/converters.go
@@ -6,13 +6,14 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/ethereum/go-ethereum/rpc"
-
-	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/obscuronet/go-obscuro/contracts/generated/MessageBus"
 	"github.com/obscuronet/go-obscuro/go/common"
+	"github.com/obscuronet/go-obscuro/go/common/errutil"
 	"github.com/obscuronet/go-obscuro/go/common/rpc/generated"
+
+	gethcommon "github.com/ethereum/go-ethereum/common"
 )
 
 // Functions to convert classes that need to be sent between the host and the enclave to and from their equivalent
@@ -48,16 +49,6 @@ func ToBlockSubmissionResponseMsg(response *common.BlockSubmissionResponse) (gen
 	}, nil
 }
 
-func ToBlockSubmissionRejectionMsg(rejectError *common.BlockRejectError) (generated.BlockSubmissionResponseMsg, error) {
-	errMsg := &generated.BlockSubmissionErrorMsg{
-		Cause:  rejectError.Wrapped.Error(),
-		L1Head: rejectError.L1Head.Bytes(),
-	}
-	return generated.BlockSubmissionResponseMsg{
-		Error: errMsg,
-	}, nil
-}
-
 func ToSecretRespMsg(responses []*common.ProducedSecretResponse) []*generated.SecretResponseMsg {
 	respMsgs := make([]*generated.SecretResponseMsg, len(responses))
 
@@ -89,7 +80,7 @@ func FromSecretRespMsg(secretResponses []*generated.SecretResponseMsg) []*common
 
 func FromBlockSubmissionResponseMsg(msg *generated.BlockSubmissionResponseMsg) (*common.BlockSubmissionResponse, error) {
 	if msg.Error != nil {
-		return nil, &common.BlockRejectError{
+		return nil, &errutil.BlockRejectError{
 			L1Head:  gethcommon.BytesToHash(msg.Error.L1Head),
 			Wrapped: errors.New(msg.Error.Cause),
 		}
